Hinzufügen der Contracts und DTOs für das InfoScreen Modul und die Sägemaßkorrekturen
