+++ Gandalan.IDAS.Client.Contracts/Contracts/UIServices/IInfoScreenDataInitiator.cs
@@ -8,13 +8,18 @@ using System.Windows;
 namespace Gandalan.Client.Contracts.UIServices
 {
     /// <summary>
-    /// Interface für die Implementierung von Initiatoren für die InfoScreens. Der Initiator eines ist für die Beschaffung der Daten für den InfoScreen zuständig (z.B. nach QR Code Scan oder PCode Eingabe)
+    /// Interface für die Implementierung von Initiatoren für die InfoScreens. 
+    /// Der Initator liefert die Notwendigen Daten an die InfoScreen Module
+    /// Standard Initiatoren: PCodeInitiator, ScanInitiator, ScanPCodeInitiator
     /// </summary>
     public interface IInfoScreenDataInitiator
     {
         string Name { get; }
+
         List<InfoScreenInitTypeDTO> GetInitData();
+
         object GetInitControl();
+
         List<InfoScreenInitTypeDTO> GetAllowedTypes();
 
         Task UnLoadAsync();
