+++ Gandalan.IDAS.WebApi.Data/DTOs/Produktion/SaegemassKorrekturDTO.cs
@@ -1,0 +1,19 @@
+﻿using System;
+
+namespace Gandalan.IDAS.WebApi.DTO
+{
+    /// <summary>
+    /// DTO für die Verwaltung der Sägemaßkorrekturen
+    /// </summary>
+    public class SaegemassKorrekturDTO
+    {
+        public DateTime ChangedDate { get; set; }
+
+        public string KatalogNummer { get; set; }
+
+        public float Korrektur90Grad { get; set; }
+
+        public float Korrektur45Grad { get; set; }
+
+    }
+}
